isPalindrome makes copy of queue: bug! double free
isPalindrome() makes a copy of queue when it gets passed in and dequeues the copy's values instead of the original's. the tailptr for the stack gets freed twice and I don't really know why? It seems to happen when main closes so it might be trying to delete tailptr again after I already deleted it