Found the bug I think
The traversal of the tree seems to be going just fine. It seems like the problem is that when I make cursor equal to the subtree child and allocate that a new pointer the subtree doesn't actually see it. So I have to have subtree equal the cursor when I allocate it memory so it knows where to go.